// client/src/app/apis/model/apiResultLoginUser.ts
/**
 * api
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v2.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { LoginUser } from './loginUser';


export interface ApiResultLoginUser { 
    readonly data?: LoginUser;
    readonly error?: string;
    readonly result?: boolean;
}